requirements.txt
cffi==1.15.1
cryptography==38.0.4
pycparser==2.21
pycryptodome==3.15.0
pyOpenSSL==22.1.0
